+++ pysys/launcher/console_run.py
@@ -111,8 +111,7 @@ Execution options
        --ci                    set optimal options for automated/non-interactive test execution in a CI job: 
                                  --purge --record -j0 --type=auto --mode=ALL --printLogs=FAILURES -XcodeCoverage
    -v, --verbosity LEVEL       set the verbosity for most pysys logging (CRIT, WARN, INFO, DEBUG)
-                   CAT=LEVEL   set the verbosity for a specific PySys logging category e.g. -vassertions=, -vprocess=
-                               (or to set the verbosity for a non-PySys Python logger category use "python:CAT=LEVEL")
+                   CAT=LEVEL   set the verbosity for a PySys/Python logging category e.g. -vassertions=, -vprocess=
    -y, --validateOnly          test the validate() method without re-running execute()
    -h, --help                  print this message
  
@@ -241,14 +240,9 @@ e.g.
 				verbosity = value
 				if '=' in verbosity:
 					loggername, verbosity = value.split('=')
-					assert not loggername.startswith('pysys.'), 'The "pysys." prefix is assumed and should not be explicitly specified'
-					if loggername.startswith('python:'):
+					if loggername.startswith('python:'): # this is weird and undocumented, but leave it in place just in case someone is using it
 						loggername = loggername[len('python:'):]
 						assert not loggername.startswith('pysys'), 'Cannot use python: with pysys.*' # would produce a duplicate log handler
-						# in the interests of performance and simplicity we normally only add the pysys.* category 
-						logging.getLogger(loggername).addHandler(pysys.internal.initlogging.pysysLogHandler)
-					else:
-						loggername = 'pysys.'+loggername
 				else:
 					loggername = None
 				
@@ -269,11 +263,15 @@ e.g.
 					# not necessarily downgrade the root level (would make run.log less useful and break 
 					# some PrintLogs behaviour)
 					stdoutHandler.setLevel(verbosity)
-					if verbosity == logging.DEBUG: logging.getLogger('pysys').setLevel(logging.DEBUG)
+					if verbosity == logging.DEBUG: 
+						logging.getLogger('pysys').setLevel(logging.DEBUG)
+						logging.getLogger().setLevel(logging.DEBUG)
 				else:
 					# for specific level setting we need the opposite - only change stdoutHandler if we're 
 					# turning up the logging (since otherwise it wouldn't be seen) but also change the specified level
+					# make the user of "pysys." prefix optional
 					logging.getLogger(loggername).setLevel(verbosity)
+					logging.getLogger('pysys.'+loggername).setLevel(verbosity)
 				
 			elif option in ("-a", "--type"):
 				self.type = value