+++ test/correctness/PySys_internal_135/run.py
@@ -11,11 +11,15 @@ from pysysinternalhelpers import *
 class PySysTest(BaseTest):
 
 	def execute(self):
-		runPySys(self, self.output+'/pysys-info', ['run', '-vpython:myloggercat=info'], workingDir=self.input)
-		runPySys(self, self.output+'/pysys-debug', ['run', '-vpython:myloggercat=debug'], workingDir=self.input)
+		runPySys(self, self.output+'/pysys-info', ['run', '-vpython:myloggercat=info', '-o', self.output+'/output-info'], workingDir=self.input)
+		runPySys(self, self.output+'/pysys-debug', ['run', '-vmyloggercat=debug', '-o', self.output+'/output-debug'], workingDir=self.input)
 
 	def validate(self):
 		self.logFileContents('pysys-debug.out', includes=['Hello at .*'])
 		self.assertGrep('pysys-debug.out', expr='Hello at debug')
 		self.assertGrep('pysys-info.out', expr='Hello at debug', contains=False)
 		self.assertGrep('pysys-info.out', expr='Hello at info')
+
+		self.assertGrep('output-info/PySysTest/run.log', expr='Hello at info')
+		self.assertGrep('output-info/PySysTest/run.log', expr='Hello at debug', contains=False)
+		self.assertGrep('output-debug/PySysTest/run.log', expr='Hello at debug')